Add support for changing dataytpe in DeepAREstimator

+++ src/gluonts/distribution/laplace.py
@@ -79,7 +79,7 @@ class Laplace(Distribution):
 
         def s(mu: Tensor, b: Tensor) -> Tensor:
             ones = mu.ones_like()
-            x = F.random.uniform(-0.5 * ones, 0.5 * ones)
+            x = F.random.uniform(-0.5 * ones, 0.5 * ones, dtype=ones.dtype)
             laplace_samples = mu - b * F.sign(x) * F.log(
                 (1.0 - 2.0 * F.abs(x)).clip(1.0e-30, 1.0e30)
                 # 1.0 - 2.0 * F.abs(x)
